Fury compendium (#951)
* Added Fury Subclasses

Added the three Fury Subclasses from the Heroes book, with description and source information.

* First Level Fury Features and Abilities

Compendium update that has all of the Fury's first level Feature and Abilities, all incorporated into The Fury class's advancements, drag and drop into a character sheet will let you now choose The Fury's Subclass, Level 1 Signature Ability, Level 1 3-Ferocity Skill, and Level 1 5-Ferocity Skill.

Note: The Beast Shape feature still needs the stormwight kit to be made before that feature is fully complete.

* Fury Compendium Update and Fixes

Addressing errors and fixes from the previous Push.

* Fury Level 1 Advancements organization and Stormwight Kits

Just reorganizing and adding descriptors for the Fury's advancements to be more in line with how it looks for the Elementalist (and other classes).

Stormwight now has kits and abilities associated with those kits.

* Added missing ability

* Fury 2nd Level Compendium Features and Abilities

For features that had an effect based off your Stats, I just defaulted the effect to +2 for now but feel free to edit or remove that if you think it's a bad idea.

* Fixed up applied effects, incremented revision on fury class & subclasses

---------

Co-authored-by: Joseph <JPMeehan5@Gmail.com>