#761 - Dwarf Ancestry Update (#1097)
Updated two Dwarven Ancestries:
Grounded
Added effect description so players will know what the effect is doing.

Spark Off Your Skin
Added effect description so players will know what the effect is doing.
Disabled effect suspension and toggled the effect on, so it will be active when applied to the character instead of needing to be toggled. This should hopefully help prevent confusion with other effects that are on by default with other races.