Career Perk Choices (#1219)
Added Perk selection and updated icons and text for Perk selection under all careers.

Issue: Odd behavior in Agent where perk selection is not giving it as Intrigue in the json text even though that is selected and even reselected in an attempt to correct. Other careers don't have this problem.

Observation: While I have selected the Perk selection for a group of perks, there is no active selection. Is the selection meant to pull from a list of these Perks? Or do we still need to add every perk of that type in the list below? This would apply for both Class perks and Career perk selection for advancement.