Fury level 4 and 5 (#1253)
* Fury level 4 and 5

Added:
Level 4: Features
Level 5: Aspect Features and 9 Ferocity Abilities

* fury-level-4-5

Changes:
Debilitating Strike Ability - Added a custom effect for “Slowed.” Added an “Other” PRE to keep the description of the effects consistent with the power roll tier.

To Stone! Ability - Added a custom effect for “Slowed” and ‘Restrained’ and added the enricher for restrained in the effect description. Added an “Other” PRE to keep the effect descriptions consistent with the power roll tier.

Stormborn Feature - Added an item grant advancement.

* Review adjustments

* Changelog

---------

Co-authored-by: Joseph <JPMeehan5@Gmail.com>